Migrate to a class based view
